refactor: Pull write-to-stream out into its own file
