+++ lib/File-builder.js
@@ -9,7 +9,7 @@ const fs = require('fs')
 const csvStringify = require('csv-string').stringify
 const Transformer = require('./Transformer')
 const upath = require('upath')
-const _ = require('lodash')
+const writeToStream = require('./write-to-stream')
 
 class FileBuilder {
   constructor (options) {
@@ -174,17 +174,5 @@ function findMatchingKey (keys, value) {
     }
   }
 } // findMatchingKey
-
-async function writeToStream(stream, output) {
-  return new Promise((resolve, reject) => {
-    stream.write(
-      output,
-      err => {
-        if (err) return reject(err)
-        resolve()
-      }
-    )
-  })
-} // writeToStream
 
 module.exports = FileBuilder
